add cash on delivery options

+++ database/migrations/2023_01_07_053323_create_discounts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->longText('description');
-            $table->decimal('discount_percentage');
+            $table->bigInteger('discount_percentage');
             $table->boolean('active')->default(false);
             $table->timestamps();
         });
